+++ crates/core/src/thermostats.rs
@@ -28,8 +28,8 @@ impl Thermostat for NullThermostat {}
 /// Berendsen weak coupling thermostat.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Berendsen {
-    target: f32,
-    tau: f32,
+    target: f64,
+    tau: f64,
 }
 
 impl Berendsen {
@@ -39,7 +39,7 @@ impl Berendsen {
     ///
     /// * `target` - Target temperature (Kelvin)
     /// * `tau` - Timestep of the thermostat expressed as a multiple of the integrator's timestep
-    pub fn new(target: f32, tau: f32) -> Berendsen {
+    pub fn new(target: f64, tau: f64) -> Berendsen {
         Berendsen { target, tau }
     }
 }
@@ -48,26 +48,26 @@ impl Berendsen {
 impl Thermostat for Berendsen {
     fn post_integrate(&mut self, system: &mut System) {
         let temperature = Temperature.calculate_intrinsic(system);
-        let factor = f32::sqrt(1.0 + (self.target / temperature - 1.0) / self.tau);
+        let factor = f64::sqrt(1.0 + (self.target / temperature - 1.0) / self.tau);
 
         // !!! this block is more efficient without `par_iter`
         system.velocities = system
             .velocities
             .iter()
             .map(|&v| v * factor)
-            .collect::<Vec<Vector3<f32>>>();
+            .collect::<Vec<Vector3<f64>>>();
     }
 }
 
 /// Nose-Hoover style thermostat.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct NoseHoover {
-    target: f32,
-    freq: f32,
-    timestep: f32,
-    psi: f32,
-    factor: f32,
-    temperature: f32,
+    target: f64,
+    freq: f64,
+    timestep: f64,
+    psi: f64,
+    factor: f64,
+    temperature: f64,
 }
 
 impl NoseHoover {
@@ -78,14 +78,14 @@ impl NoseHoover {
     /// * `target` - Target temperature (Kelvin)
     /// * `freq` - Damping frequency
     /// * `timestep` - Simulation timestep
-    pub fn new(target: f32, freq: f32, timestep: f32) -> NoseHoover {
+    pub fn new(target: f64, freq: f64, timestep: f64) -> NoseHoover {
         NoseHoover {
             target,
             freq,
             timestep,
-            psi: 0 as f32,
-            factor: 0 as f32,
-            temperature: 0 as f32,
+            psi: 0 as f64,
+            factor: 0 as f64,
+            temperature: 0 as f64,
         }
     }
 }
@@ -100,14 +100,14 @@ impl Thermostat for NoseHoover {
         let dt = self.timestep;
         let psidot = self.freq.powi(2) * ((self.temperature / self.target) - 1.0);
         self.psi += psidot * (dt / 2.0);
-        self.factor = f32::exp(-self.psi * (dt / 2.0));
+        self.factor = f64::exp(-self.psi * (dt / 2.0));
 
         // !!! this block is more efficient without `par_iter`
         system.velocities = system
             .velocities
             .iter()
             .map(|&v| v * self.factor)
-            .collect::<Vec<Vector3<f32>>>();
+            .collect::<Vec<Vector3<f64>>>();
     }
 
     fn post_integrate(&mut self, system: &mut System) {
